libxl: Add suppress-vmdesc to QEMU -machine pc options
The device model state saved by QMP xen-save-devices-state doesn't
include the vmdesc json.  When restoring an HVM, xen-load-devices-state
always triggers "Expected vmdescription section, but got 0".  This is
not a problem when restore comes from a file.  However, when QEMU runs
in a linux stubdom and comes over a console, EOF is not received.  This
causes a delay restoring - though it does restore.

Setting suppress-vmdesc skips looking for the vmdesc during restore and
avoids the wait.

This is a libxl change for the non-xenfv case to match the xenfv change
made in qemu

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>
Message-Id: <20201019200050.103360-1-jandryuk@gmail.com>