x86/pv: Drop stale comment in dom0_construct_pv()
This comment has been around since c/s 1372bca0615 in 2004.  It is stale, as
it predates the introduction of struct vcpu.

It is not obvious that it was even correct at the time.  Where a vcpu (domain
at the time) has been configured to run is unrelated to construct the domain's
initial pagetables, etc.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20201029140041.18343-1-andrew.cooper3@citrix.com>
Acked-by: Jan Beulich <jbeulich@suse.com>