xen/pci: Move x86 specific code to x86 directory.
passthrough/pci.c file is common for all architecture, but there is x86
specific code in this file.

Move x86 specific code to the drivers/passthrough/io.c file to avoid
compilation error for other architecture.

As drivers/passthrough/io.c is compiled only for x86 move it to
x86 directory and rename it to hvm.c.

No functional change.

Signed-off-by: Rahul Singh <rahul.singh@arm.com>
Message-Id: <a84005e5aa6733043e043b015cde4983719c8535.1605527997.git.rahul.singh@arm.com>